refactor(groomer): 미용사 견적 apps 폴더 구조 수정
