+++ apps/groomer/src/pages/estimate/detail/index.styles.ts
@@ -1,0 +1,22 @@
+import { theme } from '@daengle/design-system';
+import { css } from '@emotion/react';
+
+export const wrapper = css`
+  height: 100vh;
+`;
+
+export const sectionDivider = css`
+  width: 100%;
+  height: 8px;
+  background-color: ${theme.colors.gray100};
+  display: block;
+  margin: 0;
+`;
+
+export const requestTitle = css`
+  padding: 24px 18px;
+`;
+
+export const button = css`
+  padding: 24px 18px;
+`;
