+++ apps/groomer/src/queries/query-keys.ts
@@ -1,4 +1,6 @@
 export const QUERY_KEYS = {
   POST_KAKAO: ['POST_KAKAO'],
   POST_JOIN: ['POST_JOIN'],
+  GET_GROOMER_ESTIMATE_LIST: ['GET_GROOMER_ESTIMATE_LIST'],
+  GET_GROOMER_ESTIMATE_DETAIL: ['GET_GROOMER_ESTIMATE_DETAIL'],
 };
