+++ apps/groomer/tsconfig.json
@@ -4,7 +4,6 @@
     "plugins": [{ "name": "next" }],
     "baseUrl": ".",
     "paths": {
-      "@services/*": ["../../packages/services/src/*"],
       "~/*": ["src/*"]
     }
   },
